TSF-1027 sletter gammelt og flytter tester
