+++ packages/prosess-aarskvantum-oms/src/ÅrskvantumIndex.tsx
@@ -1,12 +1,10 @@
 import React, { FunctionComponent } from 'react';
 import { createIntlCache, createIntl, RawIntlProvider } from 'react-intl';
-import { VerticalSpacer } from '@fpsak-frontend/shared-components/index';
 import { Behandling, KodeverkMedNavn } from '@k9-sak-web/types';
 import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
 import Aksjonspunkt from '@k9-sak-web/types/src/aksjonspunktTsType';
 import messages from '../i18n/nb_NO.json';
 import ÅrskvantumForbrukteDager from './dto/ÅrskvantumForbrukteDager';
-import Årskvantum from './components/Årskvantum';
 import Uttaksplan from './components/Uttaksplan';
 import AksjonspunktForm from './components/AksjonspunktForm';
 import Aktivitet from './dto/Aktivitet';
@@ -42,17 +40,7 @@ const ÅrskvantumIndex: FunctionComponent<ÅrsakvantumIndexProps> = ({
   submitCallback,
   aksjonspunkterForSteg = [],
 }) => {
-  const {
-    totaltAntallDager,
-    antallKoronadager,
-    restdager,
-    restTid,
-    forbrukteDager,
-    forbruktTid,
-    antallDagerArbeidsgiverDekker,
-    antallDagerInfotrygd = 0,
-    sisteUttaksplan,
-  } = årskvantum;
+  const { sisteUttaksplan } = årskvantum;
   const aktivitetsstatuser = alleKodeverk[kodeverkTyper.AKTIVITET_STATUS];
 
   return (
@@ -67,19 +55,6 @@ const ÅrskvantumIndex: FunctionComponent<ÅrsakvantumIndexProps> = ({
           isAksjonspunktOpen={isAksjonspunktOpen}
         />
       )}
-      <Årskvantum
-        totaltAntallDager={totaltAntallDager}
-        antallKoronadager={antallKoronadager}
-        restdager={restdager}
-        restTid={restTid}
-        forbrukteDager={forbrukteDager}
-        forbruktTid={forbruktTid}
-        antallDagerArbeidsgiverDekker={antallDagerArbeidsgiverDekker}
-        antallDagerInfotrygd={antallDagerInfotrygd}
-        benyttetRammemelding={sisteUttaksplan.benyttetRammemelding}
-        uttaksperioder={sisteUttaksplan.aktiviteter.flatMap(({ uttaksperioder }) => uttaksperioder)}
-      />
-      <VerticalSpacer sixteenPx />
       <Uttaksplan
         aktiviteterBehandling={sisteUttaksplan.aktiviteter}
         aktiviteterHittilIÅr={fullUttaksplan?.aktiviteter}
